refactor: BettingMoney 내부 모든 변수 BigDecimal로 변환

+++ src/main/java/domain/blackjack/Result.java
@@ -2,17 +2,19 @@ package domain.blackjack;
 
 import domain.player.BettingMoney;
 
+import java.math.BigDecimal;
+
 public enum Result {
-    BLACKJACK("승(블랙잭)", 2.5),
-    WIN("승", 2.0),
-    DRAW("무", 1.0),
-    LOSE("패", 0.0),
+    BLACKJACK("승(블랙잭)", BigDecimal.valueOf(2.5)),
+    WIN("승", BigDecimal.valueOf(2.0)),
+    DRAW("무", BigDecimal.valueOf(1.0)),
+    LOSE("패", BigDecimal.valueOf(0.0)),
     ;
 
     private final String value;
-    private final double ratio;
+    private final BigDecimal ratio;
 
-    Result(String value, double ratio) {
+    Result(String value, BigDecimal ratio) {
         this.value = value;
         this.ratio = ratio;
     }
@@ -37,7 +39,7 @@ public enum Result {
         return value;
     }
 
-    public double getRatio() {
+    public BigDecimal getRatio() {
         return ratio;
     }
 }
